# app/static/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "description": "Auth server for TutuboApp",
    "version": "1.0.0",
    "title": "Auth Server"
  },
  "tags": [
    {
      "name": "users",
      "description": "Operations about user"
    }
  ],
  "paths": {
    "/users/register": {
      "post": {
        "tags": [
          "users"
        ],
        "summary": "Register user",
        "operationId": "createUser",
        "parameters": [
          {
            "in": "header",
            "name": "x-api-key",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "description": "Created user object",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string",
                    "example": "olifer97@address.com"
                  },
                  "username": {
                    "type": "string",
                    "example": "olifer97"
                  },
                  "password": {
                    "type": "string",
                    "example": "123456"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "integer",
                      "example": 1
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Insufficient or invalid fields",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BadRequestResponse"
                }
              }
            }
          },
          "409": {
            "description": "User already registered",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConflictResponse"
                }
              }
            }
          }
        }
      }
    },
    "/users/login": {
      "post": {
        "tags": [
          "users"
        ],
        "summary": "Logs user into the system",
        "description": "",
        "operationId": "loginUser",
        "parameters": [
          {
            "in": "header",
            "name": "x-api-key",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "description": "Login credentials",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string",
                    "example": "olifer97@address.com"
                  },
                  "password": {
                    "type": "string",
                    "example": "123456"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User logged in",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "token": {
                      "type": "string",
                      "example": "aToken"
                    },
                    "user": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "integer",
                          "example": 1
                        },
                        "username": {
                          "type": "string",
                          "example": "olifer97"
                        },
                        "email": {
                          "type": "string",
                          "example": "olifer97@address.com"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Insufficient fields",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BadRequestResponse"
                }
              }
            }
          },
          "401": {
            "description": "Bad credentials or inexistent user or user blocked",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnauthorizedResponse"
                }
              }
            }
          }
        }
      }
    },
    "/users/oauth2login": {
      "post": {
        "tags": [
          "users"
        ],
        "summary": "Logs user into the system by external service",
        "description": "",
        "operationId": "oauthLogin",
        "parameters": [
          {
            "in": "header",
            "name": "x-api-key",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "description": "Login credentials",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "idToken": {
                    "type": "string",
                    "example": "someGoogleToken"
                  },
                  "photoURL": {
                    "type": "string",
                    "example": "someURL"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User logged in",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "token": {
                      "type": "string",
                      "example": "aToken"
                    },
                    "user": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "integer",
                          "example": 1
                        },
                        "username": {
                          "type": "string",
                          "example": "olifer97"
                        },
                        "email": {
                          "type": "string",
                          "example": "olifer97@address.com"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Insufficient fields",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BadRequestResponse"
                }
              }
            }
          },
          "401": {
            "description": "Bad credentials or inexistent user",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnauthorizedResponse"
                }
              }
            }
          }
        }
      }
    },
    "/users/logout": {
      "post": {
        "tags": [
          "users"
        ],
        "summary": "Logs out current logged in user session",
        "description": "",
        "operationId": "logoutUser",
        "parameters": [
          {
            "in": "header",
            "name": "access-token",
            "schema": {
              "type": "string"
            },
            "required": true
          },
          {
            "in": "header",
            "name": "x-api-key",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "205": {
            "description": "Logged out"
          },
          "401": {
            "description": "Token not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnauthorizedResponse"
                }
              }
            }
          }
        }
      }
    },
    "/users/authorize": {
      "post": {
        "tags": [
          "users"
        ],
        "summary": "Check if token is authorize",
        "description": "",
        "operationId": "authorize",
        "parameters": [
          {
            "in": "header",
            "name": "access-token",
            "schema": {
              "type": "string"
            },
            "required": true
          },
          {
            "in": "header",
            "name": "x-api-key",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Authorized",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "user": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "integer",
                          "example": 1
                        },
                        "username": {
                          "type": "string",
                          "example": "olifer97"
                        },
                        "email": {
                          "type": "string",
                          "example": "olifer97@address.com"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Token not found or invalid or user blocked",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnauthorizedResponse"
                }
              }
            }
          }
        }
      }
    },
    "/users": {
      "get": {
        "tags": [
          "users"
        ],
        "summary": "Get all users",
        "description": "",
        "operationId": "getUsers",
        "parameters": [
          {
            "in": "header",
            "name": "x-api-key",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Users Profile",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/User"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/users/{id}": {
      "get": {
        "tags": [
          "users"
        ],
        "summary": "Get user by id",
        "description": "",
        "operationId": "getUserById",
        "parameters": [
          {
            "in": "header",
            "name": "x-api-key",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User Profile",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "404": {
            "description": "Could not find user",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotFoundResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "users"
        ],
        "summary": "Edit user info",
        "description": "",
        "operationId": "editUserById",
        "parameters": [
          {
            "in": "header",
            "name": "x-api-key",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          }
        ],
        "requestBody": {
          "description": "New field values",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "picture": {
                    "type": "string",
                    "example": "images.com/myProfilePic"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User Profile",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "404": {
            "description": "Could not find user",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotFoundResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "users"
        ],
        "summary": "Delete user by id",
        "description": "",
        "operationId": "deleteUserById",
        "parameters": [
          {
            "in": "header",
            "name": "x-api-key",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          }
        ],
        "responses": {
          "204": {
            "description": "User deleted successfully"
          },
          "404": {
            "description": "Could not find user",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotFoundResponse"
                }
              }
            }
          }
        }
      }
    },
    "/users/{id}/blocked": {
      "post": {
        "tags": [
          "users"
        ],
        "summary": "Block user by id",
        "description": "",
        "operationId": "blockUserById",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          }
        ],
        "responses": {
          "204": {
            "description": "User blocked"
          },
          "404": {
            "description": "Could not find user",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotFoundResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "users"
        ],
        "summary": "Unblock user by id",
        "description": "",
        "operationId": "unblockUserById",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          }
        ],
        "responses": {
          "204": {
            "description": "User unblocked"
          },
          "404": {
            "description": "Could not find user",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotFoundResponse"
                }
              }
            }
          }
        }
      }
    },
    "/users/reset_password": {
      "post": {
        "tags": [
          "users"
        ],
        "summary": "Reset password",
        "description": "",
        "operationId": "reset_password",
        "parameters": [
          {
            "in": "header",
            "name": "x-api-key",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "description": "Email to send code",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string",
                    "example": "olifer97@address.com"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Email sent"
          },
          "400": {
            "description": "Missing fields"
          }
        }
      }
    },
    "/users/password": {
      "get": {
        "tags": [
          "users"
        ],
        "summary": "Validate code",
        "description": "",
        "operationId": "check_reset_code",
        "parameters": [
          {
            "in": "header",
            "name": "x-api-key",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "code",
            "required": true,
            "schema": {
              "type": "string",
              "example": 1456
            }
          },
          {
            "in": "query",
            "name": "email",
            "required": true,
            "schema": {
              "type": "string",
              "example": "olifer97@hostname.com"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Valid code"
          },
          "401": {
            "description": "Invalid code or email"
          }
        }
      },
      "post": {
        "tags": [
          "users"
        ],
        "summary": "Change password",
        "description": "",
        "operationId": "new_password",
        "parameters": [
          {
            "in": "header",
            "name": "x-api-key",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "code",
            "required": true,
            "schema": {
              "type": "string",
              "example": 1456
            }
          },
          {
            "in": "query",
            "name": "email",
            "required": true,
            "schema": {
              "type": "string",
              "example": "olifer97@hostname.com"
            }
          }
        ],
        "requestBody": {
          "description": "Password to change",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "password": {
                    "type": "string",
                    "example": "newpassword"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Password changed"
          },
          "400": {
            "description": "Missing fields"
          },
          "401": {
            "description": "Invalid code or email"
          }
        }
      }
    },
    "/stats": {
      "get": {
        "tags": [
          "users"
        ],
        "summary": "Get number of users through time",
        "description": "",
        "operationId": "getStats",
        "parameters": [
          {
            "in": "query",
            "name": "initial_date",
            "required": true,
            "schema": {
              "type": "string",
              "example": "06/29/20 18:03:31"
            }
          },
          {
            "in": "query",
            "name": "final_date",
            "required": true,
            "schema": {
              "type": "string",
              "example": "06/30/20 18:03:31"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Stats",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UserStats"
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "example": 1
          },
          "username": {
            "type": "string",
            "example": "olifer97"
          },
          "email": {
            "type": "string",
            "example": "olifer97@address.com"
          },
          "profile": {
            "type": "object",
            "example": {
              "picture": "url.com/image"
            }
          }
        }
      },
      "UnauthorizedResponse": {
        "type": "object",
        "properties": {
          "reason": {
            "type": "string",
            "example": "Invalid data"
          }
        }
      },
      "BadRequestResponse": {
        "type": "object",
        "properties": {
          "reason": {
            "type": "string",
            "example": "Invalid request"
          }
        }
      },
      "NotFoundResponse": {
        "type": "object",
        "properties": {
          "reason": {
            "type": "string",
            "example": "Not Found"
          }
        }
      },
      "ConflictResponse": {
        "type": "object",
        "properties": {
          "reason": {
            "type": "string",
            "example": "Already taken"
          }
        }
      },
      "UserStats": {
        "type": "object",
        "properties": {
          "count": {
            "type": "integer",
            "example": 1
          },
          "date": {
            "type": "string",
            "example": "06/29/20 18:03:31"
          }
        }
      }
    }
  }
}